Feat/#1 컴포넌트 추가 및 과제
